added BaseModel.References to track all PuckPicker references in a model. this means its possible to see which other content a content item requires/is linked to. so if you have an image linked to the homepage you will be able to query if that image is being referred to by other content. this helps when you want to delete images or other content and want to first check that it isn't being used elsewhere.
though the initial logic for this is now complete, it will still require an EditorTemplate to show references on the edit form. i've added PuckReferences.cshtml but it's empty until i implement it. it will call a controller and query for any content with References[0].Id AND References[0].Variant equal to the model being edited